UI Improvements and Turkish Language added

+++ app/src/main/java/com/lb/apkparserdemo/apk_info/app_icon/AppInfoUtil.kt
@@ -13,7 +13,7 @@ fun PackageInfo.isSystemApp() = this.applicationInfo.isSystemApp()
 fun PackageManager.getInstalledPackagesCompat(flags: Int = 0): MutableList<PackageInfo> {
     if (VERSION.SDK_INT >= VERSION_CODES.TIRAMISU)
         return getInstalledPackages(PackageManager.PackageInfoFlags.of(flags.toLong()))
-    return getInstalledPackages(flags.toInt())
+    return getInstalledPackages(flags)
 }
 
 object AppInfoUtil {
